+++ src/main/java/chess/game/Score.java
@@ -4,13 +4,8 @@ import chess.chessboard.Board;
 import chess.chessboard.position.File;
 import chess.chessboard.position.Position;
 import chess.chessboard.position.Rank;
-import game.Player;
 
 import java.util.HashMap;
-
-import static game.Player.BLACK;
-import static game.Player.WHITE;
-
 
 public final class Score {
 
@@ -25,21 +20,21 @@ public final class Score {
     }
 
     public void calculateScore(final Board board) {
-        scores.put(WHITE, calculateScore(board, WHITE));
-        scores.put(BLACK, calculateScore(board, BLACK));
+        scores.put(Player.WHITE, calculateScore(board, Player.WHITE));
+        scores.put(Player.BLACK, calculateScore(board, Player.BLACK));
     }
 
     private double calculateScore(final Board board, final Player player) {
         double score = STARTING_COUNT;
-        for (Rank rank : Rank.values()) {
-            score = scoresInRank(board, player, score, rank);
+        for (File file : File.values()) {
+            score = scoresInFile(board, player, score, file);
         }
         return score;
     }
 
-    private double scoresInRank(final Board board, final Player player, double score, final Rank rank) {
+    private double scoresInFile(final Board board, final Player player, double score, final File file) {
         int pawnCountInFile = STARTING_COUNT;
-        for (File file : File.values()) {
+        for (Rank rank : Rank.values()) {
             score = addScoreOfPiece(board, player, score, Position.of(rank, file));
             pawnCountInFile += addPawnCount(board, player, Position.of(rank, file));
         }
@@ -50,7 +45,7 @@ public final class Score {
     }
 
     private int addPawnCount(final Board board, final Player player, final Position position) {
-        if (board.isSamePlayerIn(position, player) || board.isPawn(position)) {
+        if (board.isSamePlayerIn(position, player) && board.isPawn(position)) {
             return COUNTS;
         }
         return NONE_SUM;
