+++ src/main/java/chess/piece/Blank.java
@@ -1,6 +1,6 @@
 package chess.piece;
 
-import game.Player;
+import chess.game.Player;
 import chess.chessboard.position.Position;
 
 import java.util.Map;
