+++ src/main/java/chess/state/Started.java
@@ -3,7 +3,7 @@ package chess.state;
 import chess.chessboard.Board;
 import chess.chessboard.position.Position;
 import chess.piece.Piece;
-import game.Player;
+import chess.game.Player;
 
 import java.util.Map;
